+++ src/main.js
@@ -1,29 +1,30 @@
 'use strict';
 
 
-require('./form');
-require('./game');
-require('./reviews');
+var form = require('./form');
+var Game = require('./game');
+var reviewData = require('./reviews');
 
-(function() {
-  var game = new window.Game(document.querySelector('.demo'));
-  game.initializeLevelAndStart();
-  game.setGameStatus(window.Game.Verdict.INTRO);
 
+  var game = new Game(document.querySelector('.demo'));
   var formOpenButton = document.querySelector('.reviews-controls-new');
 
   /** @param {MouseEvent} evt */
   formOpenButton.onclick = function(evt) {
     evt.preventDefault();
 
-    window.form.open(function() {
-      game.setGameStatus(window.Game.Verdict.PAUSE);
+      form.open(function() {
+      game.setGameStatus(Game.Verdict.PAUSE);
       game.setDeactivated(true);
     });
   };
 
-  window.form.onClose = function() {
+    form.onClose = function() {
     game.setDeactivated(false);
   };
-  window.reviews.load();
-})();
+
+  game.initializeLevelAndStart();
+  game.setGameStatus(Game.Verdict.INTRO);
+
+  reviewData();
+
